+++ standalone/include/psen_scan_v2_standalone/configuration/default_parameters.h
@@ -30,14 +30,15 @@ static constexpr unsigned short CONTROL_PORT_OF_SCANNER_DEVICE{ 3000 };
 static constexpr unsigned short DATA_PORT_OF_HOST_DEVICE{ 55115 };
 static constexpr unsigned short CONTROL_PORT_OF_HOST_DEVICE{ 55116 };
 
-static constexpr bool DEFAULT_FRAGMENTED_SCANS{ false };
+static constexpr bool FRAGMENTED_SCANS{ false };
+static constexpr bool INTENSITIES{ false };
+static constexpr bool DIAGNOSTICS{ false };
 
 //! @brief Start angle of measurement.
-static constexpr double DEFAULT_ANGLE_START(-data_conversion_layer::degreeToRadian(137.5));
+static constexpr double DEFAULT_ANGLE_START(-data_conversion_layer::degreeToRadian(137.4));
 //! @brief  End angle of measurement.
-static constexpr double DEFAULT_ANGLE_END(data_conversion_layer::degreeToRadian(137.5));
-
-static constexpr uint16_t NUMBER_OF_SAMPLES_FULL_SCAN_MASTER{ 2750 };
+static constexpr double DEFAULT_ANGLE_END(data_conversion_layer::degreeToRadian(137.4));
+static constexpr double DEFAULT_SCAN_ANGLE_RESOLUTION(data_conversion_layer::degreeToRadian(0.1));
 
 static constexpr double TIME_PER_SCAN_IN_S{ 0.03 };