+++ standalone/include/psen_scan_v2_standalone/protocol_layer/scanner_state_machine.h
@@ -221,7 +221,7 @@ public:  // Definition of state machine via table
 private:
   // LCOV_EXCL_START
   /**
-   * @brief This error is thrown when something goes wrong with the scanner reply.
+   * @brief Exception thrown when something goes wrong with the scanner reply.
    *
    * For example an unexpected code in reply, request refused by device or unknown operation result code.
    */