+++ standalone/include/psen_scan_v2_standalone/util/tenth_of_degree.h
@@ -65,7 +65,18 @@ public:
     return TenthOfDegree(value() * rhs);
   }
 
+  constexpr TenthOfDegree operator*(const size_t& rhs) const
+  {
+    return TenthOfDegree(value() * rhs);
+  }
+
   constexpr TenthOfDegree& operator*(const int& rhs)
+  {
+    tenth_of_degree_ = value() * rhs;
+    return *this;
+  }
+
+  constexpr TenthOfDegree& operator*(const size_t& rhs)
   {
     tenth_of_degree_ = value() * rhs;
     return *this;
@@ -102,6 +113,11 @@ public:
   constexpr bool operator==(const TenthOfDegree& rhs) const
   {
     return value() == rhs.value();
+  }
+
+  constexpr bool operator!=(const TenthOfDegree& rhs) const
+  {
+    return value() != rhs.value();
   }
 
   constexpr bool operator>=(const TenthOfDegree& rhs) const